Split off hyperbolic functions from trig fns..
Add Gudermannian from autoload Guidermannian.m
